Remove the ast transform
