--- app/helpers/-element.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-element-helper/helpers/-element';
